+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### v1.5
 - Added ability to pick up XP orbs (Thanks Bletch1971!)
+- Added configuration option to turn off picking up XP orbs
+- Added configuration option to modify the range of blocks around the XP Block in which to pickup XP orbs
 - Fixed configuration option for comparator output not working properly. The config option has been moved to server config (found in the serverconfig folder in the world folder)
 
 #### v1.4.15